Build off of stable/liberty
There were upstream changes to keystone's wsgi file between liberty and mitaka.
This forced us to pass in a SHA to build off of instead of a branch, like
stable/liberty. Being able to build off of a branch makes more sense and is
easier to maintain.

With an upstream role change [0], we can deploy using branches.

[0] https://github.com/os-cloud/openstack-ansible-os_keystone/blob/master/templates/keystone-wsgi.py.j2

Change-Id: I67cd03f91fc15baafba6154473f06e0ef4a8977d