Add a really simple gate check
Right now the capstone-deploy project doesn't really gate on anything after
capstone is setup. This commit just adds a curl statement after the
capstone-deploy ansible bits finish to check that keystone is at least up and
running.

This type of gating will be hardened in the future with integration testing
that is dependent on capstone and ansible dependencies (os_keystone).

Change-Id: I751eafe457de5db3bb27271d42b1933dd16dc448