Changed Keystone repo to use the OpenStack repo
This change updates the keystone role repo to use the OpenStack role we've moved
into the OpenStack namespace. The change will ensure you're able to pull in the latest
updates the OpenStack Ansible community works on while also giving people within the
OpenStack community access to work on the role via Gerrit.